Add no_ancient_experiments rule, closes #25
For now the actual list of ancient experiments is not residing
in the repo (for NDA reasons). If you want to test the functionality
on local machine, run

    npm install
    npm run generate-ancients

In future, when we move the project to devtools.git, we can bundle
the list into linter and sublime package in a release script. Probably
the best time to do it is during hackaton.